Fix TLS alignment issue on x86_64
According to the x86_64 ELF ABI, the thread pointer in FS register
should be aligned according to the alignment of the TLS program header
in the ELF file. As this pointer should be just after the TLS area, the
TLS area itself has to be padded at the front to ensure that alignment.